feat: 어뷰징 방지 4종 구현(CLAUDE.md §6) + 어드민 모니터링 페이지 (#36)
CLAUDE.md §6에 명시됐지만 전부 미구현이던 4개 항목을 실시간(댓글
작성/투표/추천 시점 인라인 처리, 별도 배치/스케줄러 없음)으로 구현:

1. 동일 IP 해시 다수 익명 ID 감지 → 신뢰도 가중치 하향
   abuse/TrustScoreService.java — anon_users.ip_hash_recent로 클러스터
   조회, 클러스터 크기가 임계값(기본 4) 이상이면 전원(새로 걸린
   계정만이 아니라 클러스터 전체) trust_score에 페널티 배수 적용.
   신뢰도는 증분 가감이 아니라 호출마다 전체 재계산(가입 후 경과시간
   기반 성숙도 × 클러스터 페널티) — 순서 의존 드리프트 방지.
   admin_alerts에 multi_account_same_ip 기록(24시간 내 중복 알림
   방지, RateLimiter와 동일한 Redis increment+TTL 패턴을 setIfAbsent
   마커로 응용).

2. 동일/유사 반복 댓글 자동 필터
   moderation/DuplicateCommentFilter.java(레벤슈타인 직접 구현 —
   commons-lang3의 동명 메서드는 deprecated) + abuse/DuplicateCommentGuard.java.
   같은 익명 ID의 최근 댓글(기본 10개, 24시간 이내)과 유사도(기본
   0.85) 이상이면 기존 금칙어 필터와 동일한 400 응답으로 거부.

3. 급증 패턴 탐지 → 어드민 알림
   abuse/SpikeDetector.java — RateLimiter의 Redis increment+
   expire-on-first-hit 패턴 재사용. 투표(쌍 단위)·댓글 추천(up만,
   댓글 단위) 카운트가 시간 윈도(기본 10분) 내 임계값에 "정확히
   도달한 시점"에만 activity_spike 알림 1건(초과분마다 스팸 안 되게).

4. 추천/투표 집계 시 신뢰도 가중치 적용
   VoteRepository에 Vote-AnonUser theta-join(둘 다 UUID 컬럼일 뿐
   JPA 연관관계 없음) 가중합 쿼리 추가, PairService.getPair()가
   COUNT 대신 SUM(trustScore) 사용 — TopicPairDto의 투표 필드가
   long→double로 바뀌지만 VoteWidget은 퍼센트만 계산해써서 프론트
   변경 불필요. 댓글 추천(up/down)은 비정규화 카운터 구조라 가중치
   적용을 이번 범위에서 명시적으로 제외(후속 작업).

**죽은 컬럼 2개를 처음 실사용**: anon_users.trust_score(생성자에서
0.5로만 찍히고 어디서도 안 읽히던 컬럼), admin_alerts.resolved(항상
false로만 남아있던 컬럼) — AdminAlert.resolve() 추가.

**전부 어드민에서 조작 가능**(사용자 명시 원칙): crawl_settings/
moderation_settings와 동일한 싱글톤 행 패턴으로 abuse_settings
신설(V8 마이그레이션, 클러스터 크기/페널티 배수/성숙 기간/유사도
임계값/급증 윈도·임계값 등 10개 값). 새 /admin/abuse 페이지 —
미해결 알림(해결 처리 버튼)/IP 클러스터 현황/해결 이력 3섹션.

로컬 실 스택(postgres+redis) 검증: 동일 IP(로컬호스트)로 익명ID
3개 생성해 클러스터 임계값 넘기기 → trust_score 실제 하락(0.09 =
성숙도0.3×페널티0.3) + multi_account_same_ip 알림 확인 → 유사
댓글 재작성 시 400 거부 확인, 무관한 댓글은 통과 확인 → DB에
trust_score 1.0/0.3인 두 익명ID로 직접 투표 시딩해 GET /api/pairs/{id}가
정확히 1.0/0.3으로 가중합 반환하는 것 확인 → 투표 5회로 급증 임계값
넘겨 activity_spike 알림 정확히 1건 확인 → 해결 처리 API로 알림이
해결됨 목록으로 이동하는 것 확인 → 어드민 설정 API로 임계값
실시간 변경되는 것 확인 → /admin/abuse, /admin/settings 브라우저
렌더링 스크린샷 확인 → 설정 기본값 복원 및 테스트 데이터 전부 정리.

+++ backend/src/main/java/com/siso/backend/pair/TopicPairDto.java
@@ -8,9 +8,9 @@ public record TopicPairDto(
         String leftStance,
         String rightStance,
         OffsetDateTime createdAt,
-        long leftVotes,
-        long rightVotes,
-        long neutralVotes,
+        double leftVotes,
+        double rightVotes,
+        double neutralVotes,
         String myStance) {
 
     public static TopicPairDto from(TopicPair pair) {
@@ -18,7 +18,7 @@ public record TopicPairDto(
     }
 
     public static TopicPairDto from(
-            TopicPair pair, long leftVotes, long rightVotes, long neutralVotes, String myStance) {
+            TopicPair pair, double leftVotes, double rightVotes, double neutralVotes, String myStance) {
         return new TopicPairDto(
                 pair.getId(),
                 pair.getTitle(),
